third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/computed/iterable.html

<!doctype html>
<meta charset="utf-8">
<title>StylePropertyMap iterable tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#the-stylepropertymap">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<div id="target" style="--A: A; width: 10px; --C: C; transition-duration: 1s, 2s; color: red; --B: B;"></div>
<script>
'use strict';

function findInStyleMap(styleMap, property) {
  const index = [...styleMap.keys()].indexOf(property);
  if (index == -1)
    return null;
  return [...styleMap.values()][index];
}

const target = document.getElementById('target');
const styleMap = target.computedStyleMap();

test(() => {
  const expectedKeys = [...getComputedStyle(target)].sort().concat('--A', '--B', '--C');
  assert_array_equals([...styleMap], expectedKeys);
}, 'StylePropertyMap iterates properties in correct order');

test(() => {
  assert_style_value_equals(findInStyleMap(styleMap, 'width'), CSS.px(10));
}, 'StylePropertyMap iterator returns CSS properties with the correct CSSStyleValue');

test(() => {
  assert_style_value_array_equals(findInStyleMap(styleMap, 'transition-duration'), [CSS.s(1), CSS.s(2)]);
}, 'StylePropertyMap iterator returns list-valued properties with the correct CSSStyleValue');

test(() => {
  assert_style_value_equals(findInStyleMap(styleMap, '--A'), new CSSUnparsedValue('--A'));
  assert_style_value_equals(findInStyleMap(styleMap, '--B'), new CSSUnparsedValue('--B'));
  assert_style_value_equals(findInStyleMap(styleMap, '--C'), new CSSUnparsedValue('--C'));
}, 'StylePropertyMap iterator returns custom properties with the correct CSSStyleValue');

</script>
