third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/computed/iterable-expected.txt

This is a testharness.js-based test.
FAIL StylePropertyMap iterates properties in correct order assert_array_equals: lengths differ, expected 297 got 294
PASS StylePropertyMap iterator returns CSS properties with the correct CSSStyleValue
PASS StylePropertyMap iterator returns list-valued properties with the correct CSSStyleValue
FAIL StylePropertyMap iterator returns custom properties with the correct CSSStyleValue assert_equals: expected object "--A" but got null
Harness: the test ran to completion.
